# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ShopClone E-Commerce API",
    "description": "A full-featured e-commerce REST API powering the ShopClone storefront. Provides endpoints for user authentication, product catalog browsing, shopping cart management, order processing, wishlists, product reviews, and simulated payment processing.",
    "version": "1.0.0",
    "contact": {
      "name": "ShopClone API Support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/wishlist": {
      "get": {
        "tags": ["Wishlist"],
        "summary": "Get the current user's wishlist",
        "description": "Retrieves all products in the authenticated user's wishlist.",
        "operationId": "getWishlist",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved wishlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WishlistResponse"
                },
                "example": {
                  "products": [
                    {
                      "id": "prod-002",
                      "title": "Organic Cotton T-Shirt",
                      "description": "Comfortable and sustainable organic cotton t-shirt available in multiple colors.",
                      "price": 24.99,
                      "image": "https://example.com/images/tshirt.jpg",
                      "category": "clothing"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Wishlist"],
        "summary": "Add a product to the wishlist",
        "description": "Adds a product to the authenticated user's wishlist by its product ID.",
        "operationId": "addToWishlist",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToWishlistRequest"
              },
              "example": {
                "productId": "prod-002"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product successfully added to wishlist; returns the updated wishlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WishlistResponse"
                },
                "example": {
                  "products": [
                    {
                      "id": "prod-002",
                      "title": "Organic Cotton T-Shirt",
                      "description": "Comfortable and sustainable organic cotton t-shirt available in multiple colors.",
                      "price": 24.99,
                      "image": "https://example.com/images/tshirt.jpg",
                      "category": "clothing"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (missing productId)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Product ID is required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/wishlist/{productId}": {
      "delete": {
        "tags": ["Wishlist"],
        "summary": "Remove a product from the wishlist",
        "description": "Removes a specific product from the authenticated user's wishlist by its product ID.",
        "operationId": "removeFromWishlist",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the product to remove from the wishlist",
            "schema": {
              "type": "string"
            },
            "example": "prod-002"
          }
        ],
        "responses": {
          "200": {
            "description": "Product successfully removed from wishlist; returns the updated wishlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WishlistResponse"
                },
                "example": {
                  "products": []
                }
              }
            }
          },
          "400": {
            "description": "Invalid product ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid product ID"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/reviews": {
      "post": {
        "tags": ["Reviews"],
        "summary": "Submit a product review",
        "description": "Creates a new review for a product. Requires authentication. The reviewer's name is automatically derived from the authenticated user's profile.",
        "operationId": "createReview",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              },
              "example": {
                "productId": "prod-001",
                "rating": 5,
                "title": "Excellent sound quality!",
                "body": "These headphones have amazing sound quality and the noise cancellation is top-notch. Battery life is as advertised."
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleReviewResponse"
                },
                "example": {
                  "review": {
                    "id": "rev-xyz789",
                    "productId": "prod-001",
                    "userName": "Jane Doe",
                    "rating": 5,
                    "title": "Excellent sound quality!",
                    "body": "These headphones have amazing sound quality and the noise cancellation is top-notch. Battery life is as advertised.",
                    "date": "2026-02-23T14:30:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (missing required fields or invalid rating)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Product ID, rating, title, and body are required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from the /api/auth/signup or /api/auth/signin endpoints. Include in the Authorization header as: Bearer <token>"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "description": "Represents a product in the catalog",
        "required": ["id", "title", "description", "price", "image", "category"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the product",
            "example": "prod-001"
          },
          "title": {
            "type": "string",
            "description": "Name of the product",
            "example": "Wireless Bluetooth Headphones"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the product",
            "example": "Premium noise-cancelling wireless headphones with 30-hour battery life."
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Price of the product in USD",
            "minimum": 0,
            "example": 79.99
          },
          "image": {
            "type": "string",
            "format": "uri",
            "description": "URL of the product image",
            "example": "https://example.com/images/headphones.jpg"
          },
          "category": {
            "type": "string",
            "description": "Category the product belongs to",
            "example": "electronics"
          }
        }
      },
...
      "Review": {
        "type": "object",
        "description": "Represents a product review submitted by a user",
        "required": ["id", "productId", "userName", "rating", "title", "body", "date"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the review",
            "example": "rev-xyz789"
          },
          "productId": {
            "type": "string",
            "description": "The ID of the product being reviewed",
            "example": "prod-001"
          },
          "userName": {
            "type": "string",
            "description": "Display name of the reviewer",
            "example": "Jane Doe"
          },
          "rating": {
            "type": "integer",
            "description": "Rating from 1 to 5 stars",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "title": {
            "type": "string",
            "description": "Brief title summarizing the review",
            "example": "Excellent sound quality!"
          },
          "body": {
            "type": "string",
            "description": "Full text of the review",
            "example": "These headphones have amazing sound quality and the noise cancellation is top-notch. Battery life is as advertised."
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp of when the review was submitted",
            "example": "2026-02-20T10:15:00.000Z"
          }
        }
      },
...
      "AddToWishlistRequest": {
        "type": "object",
        "description": "Request body for adding a product to the wishlist",
        "required": ["productId"],
        "properties": {
          "productId": {
            "type": "string",
            "description": "The unique identifier of the product to add to the wishlist",
            "example": "prod-002"
          }
        }
      },
      "CreateReviewRequest": {
        "type": "object",
        "description": "Request body for submitting a product review",
        "required": ["productId", "rating", "title", "body"],
        "properties": {
          "productId": {
            "type": "string",
            "description": "The ID of the product being reviewed",
            "example": "prod-001"
          },
          "rating": {
            "type": "integer",
            "description": "Rating from 1 to 5 stars",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "title": {
            "type": "string",
            "description": "Brief title summarizing the review",
            "minLength": 1,
            "example": "Excellent sound quality!"
          },
          "body": {
            "type": "string",
            "description": "Full text of the review",
            "minLength": 1,
            "example": "These headphones have amazing sound quality and the noise cancellation is top-notch. Battery life is as advertised."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message describing what went wrong",
            "example": "An unexpected error occurred"
          }
        }
      },
...
      "WishlistResponse": {
        "type": "object",
        "description": "Response containing the user's wishlist products",
        "required": ["products"],
        "properties": {
          "products": {
            "type": "array",
            "description": "Array of product objects in the wishlist",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          }
        }
      },
...
      "SingleReviewResponse": {
        "type": "object",
        "description": "Response containing a single review",
        "required": ["review"],
        "properties": {
          "review": {
            "$ref": "#/components/schemas/Review"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or token is invalid/expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Authentication required"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "An unexpected server error occurred",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Internal server error"
            }
          }
        }
      }
    }
  }
}